ci: moved from MSTest to Xunit
