+++ GrowthBook.Tests/GrowthBook.Tests.csproj
@@ -1,0 +1,51 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <PlatformTarget>AnyCPU</PlatformTarget>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="Json/custom-cases.json" />
+    <None Remove="Json/FeatureDictionary.json" />
+    <None Remove="Json/GrowthBookContext.json" />
+    <None Remove="Json/GrowthBookContext.NoFeatures.json" />
+    <None Remove="Json/SingleFeatureDictionary.WithNameSpace.json" />
+    <None Remove="Json/standard-cases.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Json/custom-cases.json" />
+    <EmbeddedResource Include="Json/FeatureDictionary.json" />
+    <EmbeddedResource Include="Json/GrowthBookContext.json" />
+    <EmbeddedResource Include="Json/GrowthBookContext.NoFeatures.json" />
+    <EmbeddedResource Include="Json/SingleFeatureDictionary.WithNameSpace.json" />
+    <EmbeddedResource Include="Json/standard-cases.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="FluentAssertions" Version="6.5.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="5.0.10" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
+    <PackageReference Include="xunit" Version="2.4.1" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="XunitXml.TestLogger" Version="3.0.70" />
+    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="test-results/" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../GrowthBook/GrowthBook.csproj" />
+  </ItemGroup>
+
+</Project>
