--- Test/GrowthBook.Tests.csproj
@@ -1,41 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6</TargetFramework>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Json\custom-cases.json" />
-    <None Remove="Json\FeatureDictionary.json" />
-    <None Remove="Json\GrowthBookContext.json" />
-    <None Remove="Json\GrowthBookContext.NoFeatures.json" />
-    <None Remove="Json\SingleFeatureDictionary.WithNameSpace..json" />
-    <None Remove="Json\standard-cases.json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Json\custom-cases.json" />
-    <EmbeddedResource Include="Json\FeatureDictionary.json" />
-    <EmbeddedResource Include="Json\GrowthBookContext.json" />
-    <EmbeddedResource Include="Json\GrowthBookContext.NoFeatures.json" />
-    <EmbeddedResource Include="Json\SingleFeatureDictionary.WithNameSpace..json" />
-    <EmbeddedResource Include="Json\standard-cases.json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="MSTest" Version="3.1.1" />
-    <PackageReference Include="MSTest.TestAdapter" Version="3.1.1" />
-    <PackageReference Include="MSTest.TestFramework" Version="3.1.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GrowthBook\GrowthBook.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Properties\" />
-  </ItemGroup>
-
-</Project>
